added configurable colors for damage hud
