+++ Race_Element/Controls/Info/ReleaseNotes.cs
@@ -6,7 +6,8 @@ public static class ReleaseNotes
 {
     internal readonly static Dictionary<string, string> Notes = new()
     {
-        {"1.0.8.6", "- Internal optimizations to rendering."+
+        {"1.0.8.6", "- Damage HUD: Added configurable colors for each level of body damage"+
+                    "\n- Internal optimizations to rendering."+
                     "\n- Improved data detection for cars leaving the server by Andrei Jianu."+
                     "\n- Added hud to benchmark rendering optimizations (Pitwall->CB Benchmark)."},
         {"1.0.8.4", "- Fuel Info HUD: Added option to set which laptime is used in the fuel calculation."+
